Address reviewer feedback
